pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bav-dqs-framework"
version = "1.0.0.dev"
description = "Boundary-Aware Validity Framework for Digital Quantum Simulation"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }

authors = [
  { name = "Elionai Moura Cordeiro", email = "elionai@embits.digital" }
]

keywords = [
  "quantum-simulation",
  "lattice-field-theory",
  "zitterbewegung",
  "causal-inference",
  "quantum-computing"
]

classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Physics",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11"
]

dependencies = [
  "numpy>=1.24",
  "h5py>=3.9",
  "pyyaml>=6.0",
  "matplotlib>=3.7",
  "qiskit==2.3.0",
  "qiskit-aer==0.17.2"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "black>=23.9",
  "ruff>=0.1.5",
  "mypy>=1.6"
]

[project.urls]
Homepage = "https://github.com/embits-digital/bav-dqs-framework"
Repository = "https://github.com/embits-digital/bav-dqs-framework"
Issues = "https://github.com/embits-digital/bav-dqs-framework/issues"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["bav_dqs", "bav_dqs.*"]
namespaces = false

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
